fix(core): show the workspace name instead of the Goovee brand (#116645)
The root metadata declared a title template of `Goovee` rather than `%s`, and a
Next.js template without a placeholder overwrites the child title, so the
workspace name already computed by the workspace layout never reached the tab.

- The browser tab and the login screen now carry the workspace name; the auth
  pages, which sit outside the workspace segment, resolve it as a guest from the
  `workspaceURI` and `tenant` search params.
- `reset-password/[email]` was a client component and could not carry
  `generateMetadata`, so it is wrapped in a server page.
- Notification mails interpolate the workspace name, and their signature block
  goes through i18n instead of raw HTML.
- Everything left without a workspace context — the sidebar fallback,
  `GET /api/info`, and the OTP, invite and reset-password fallback templates used
  when no AOS template is configured — falls back to the Axelor brand.

Technical occurrences are untouched: generated imports, `GOOVEE_PUBLIC_*`
environment variables, `findGooveeUserByEmail` and the `--goovee-*` CSS
variables.
