+++ app/api/info/route.ts
@@ -1,11 +1,12 @@
 import {NextResponse} from 'next/server';
-import goovee from '@/package.json';
+import pkg from '@/package.json';
 import {manager} from '@/tenant';
+import {APP_TITLE} from '@/constants';
 
 export function GET() {
   return NextResponse.json({
-    name: 'Goovee',
-    version: goovee?.version,
+    name: APP_TITLE,
+    version: pkg?.version,
     date: new Date().toISOString(),
     tenancy: manager.getType(),
   });
