+++ lib/core/otp/actions/index.ts
@@ -8,6 +8,7 @@ import {
   replacePlaceholders,
 } from '@/orm/email-template';
 import type {Client} from '@/goovee/.generated/client';
+import {APP_TITLE} from '@/constants';
 
 // ---- LOCAL IMPORTS ---- //
 import {create} from '../orm';
@@ -29,7 +30,7 @@ const otpTemplate = ({
   otp: string;
   subject?: string;
 }) => ({
-  subject: 'Goovee OTP',
+  subject: `${APP_TITLE} OTP`,
   to: email,
   html: otpTemplateHTML({otp}),
 });
@@ -38,7 +39,7 @@ const otpTemplateHTML = ({otp}: {otp: string}) => `
     <!DOCTYPE html>
     <html>
     <head>
-        <title>Goovee OTP</title>
+        <title>${APP_TITLE} OTP</title>
         <style>
             body {
                 font-family: Arial, sans-serif;
@@ -79,12 +80,12 @@ const otpTemplateHTML = ({otp}: {otp: string}) => `
     <body>
         <div class="container">
             <div class="header">
-                <h1>OTP for Goovee</h1>
+                <h1>OTP for ${APP_TITLE}</h1>
             </div>
-            <p>Your OTP for goovee is : <strong>${otp}</strong></p>
+            <p>Your OTP for ${APP_TITLE} is : <strong>${otp}</strong></p>
             <p>Please don't share your OTP with anyone. OTP is valid for 10 minutes</p>
             <div class="footer">
-                <p>Best regards,<br>The Goovee Team</p>
+                <p>Best regards,<br>The ${APP_TITLE} Team</p>
             </div>
         </div>
     </body>
@@ -126,7 +127,7 @@ export async function generateOTP({
 
           await mailService?.notify({
             to: email,
-            subject: template?.subject || 'Greetings from Goovee',
+            subject: template?.subject || `Greetings from ${APP_TITLE}`,
             html: replacePlaceholders({
               content: template?.content,
               values: {
