+++ public/locales/en.json
@@ -141,6 +141,7 @@
   "Bank transfers may immediately use your existing Stripe balance.": "Bank transfers may immediately use your existing Stripe balance.",
   "Be the first to start a discussion in this group.": "Be the first to start a discussion in this group.",
   "Best answer": "Best answer",
+  "Best regards,": "Best regards,",
   "Billing & delivery address": "Billing & delivery address",
   "Billing address": "Billing address",
   "Brand": "Brand",
@@ -519,7 +520,6 @@
   "Go to Homepage": "Go to Homepage",
   "Go to the homepage": "Go to the homepage",
   "Good": "Good",
-  "Goovee": "Goovee",
   "Grid view": "Grid view",
   "Group": "Group",
   "Group id is missing": "Group id is missing",
@@ -942,7 +942,6 @@
   "Post reply": "Post reply",
   "Postal address": "Postal address",
   "Powered by": "Powered by",
-  "Powered by Goovee": "Powered by Goovee",
   "Preference updated": "Preference updated",
   "Preference updated successfully": "Preference updated successfully",
   "Preferences": "Preferences",
@@ -1240,6 +1239,7 @@
   "Thank you very much for your payment, you are now registered for this event": "Thank you very much for your payment, you are now registered for this event",
   "The contact you invite must be people from your company as their account will be linked to your company.": "The contact you invite must be people from your company as their account will be linked to your company.",
   "The link may be outdated or the resource has been moved. Head back to your workspace to continue.": "The link may be outdated or the resource has been moved. Head back to your workspace to continue.",
+  "The {0} Team": "The {0} Team",
   "There are no events": "There are no events",
   "There are no events today": "There are no events today",
   "These details will appear on your badge.": "These details will appear on your badge.",
@@ -1414,7 +1414,7 @@
   "You have been registered for an event!": "You have been registered for an event!",
   "You have been successfully registered to this event.": "You have been successfully registered to this event.",
   "You have blocked notifications for this site. To receive updates, please enable them in your browser settings.": "You have blocked notifications for this site. To receive updates, please enable them in your browser settings.",
-  "You have received new notification from Goovee {0}": "You have received new notification from Goovee {0}",
+  "You have received new notification from {0} {1}": "You have received new notification from {0} {1}",
   "You have unpaid invoices that will soon reach their limit date.": "You have unpaid invoices that will soon reach their limit date.",
   "You have {0} new comments on \"{1}\"": "You have {0} new comments on \"{1}\"",
   "You have {0} new replies to your comment": "You have {0} new replies to your comment",
@@ -1501,7 +1501,7 @@
   "votes": "votes",
   "with project task link type": "with project task link type",
   "z-A": "z-A",
-  "{0} - Notifications from Goovee": "{0} - Notifications from Goovee",
+  "{0} - Notifications from {1}": "{0} - Notifications from {1}",
   "{0} added a comment": "{0} added a comment",
   "{0} added a comment on {1}": "{0} added a comment on {1}",
   "{0} application(s) enabled": "{0} application(s) enabled",
