+++ public/locales/fr.json
@@ -141,6 +141,7 @@
   "Bank transfers may immediately use your existing Stripe balance.": "Les virements bancaires peuvent utiliser immédiatement votre solde Stripe existant.",
   "Be the first to start a discussion in this group.": "Soyez le premier à lancer une discussion dans ce groupe.",
   "Best answer": "Meilleure réponse",
+  "Best regards,": "Cordialement,",
   "Billing & delivery address": "Adresse de facturation et de livraison",
   "Billing address": "Adresse de facturation",
   "Brand": "Marque",
@@ -519,7 +520,6 @@
   "Go to Homepage": "Aller à la page d'accueil",
   "Go to the homepage": "Retourner à la page d'accueil",
   "Good": "Bon",
-  "Goovee": "Goovee",
   "Grid view": "Vue grille",
   "Group": "Groupe",
   "Group id is missing": "L'identifiant du groupe est manquant",
@@ -942,7 +942,6 @@
   "Post reply": "Publier la réponse",
   "Postal address": "Adresse postale",
   "Powered by": "Alimenté par",
-  "Powered by Goovee": "Propulsé par Goovee",
   "Preference updated": "Préférence mise à jour",
   "Preference updated successfully": "Préférence mise à jour avec succès",
   "Preferences": "Préférences",
@@ -1240,6 +1239,7 @@
   "Thank you very much for your payment, you are now registered for this event": "Merci pour votre paiement, vous êtes maintenant inscrit à cet événement",
   "The contact you invite must be people from your company as their account will be linked to your company.": "Le contact que vous invitez doit être une personne de votre entreprise, car son compte sera lié à votre entreprise",
   "The link may be outdated or the resource has been moved. Head back to your workspace to continue.": "Le lien est peut-être obsolète ou la ressource a été déplacée. Retournez à votre espace pour continuer.",
+  "The {0} Team": "L'équipe {0}",
   "There are no events": "Il n'y a pas d'événements",
   "There are no events today": "Il n’y a pas d’événements aujourd’hui",
   "These details will appear on your badge.": "Ces informations figureront sur votre badge.",
@@ -1414,7 +1414,7 @@
   "You have been registered for an event!": "Vous avez été inscrit à un événement !",
   "You have been successfully registered to this event.": "Vous avez été inscrit avec succès à cet événement",
   "You have blocked notifications for this site. To receive updates, please enable them in your browser settings.": "Vous avez bloqué les notifications pour ce site. Pour recevoir des mises à jour, veuillez les activer dans les paramètres de votre navigateur.",
-  "You have received new notification from Goovee {0}": "Vous avez reçu une nouvelle notification de Goovee {0}",
+  "You have received new notification from {0} {1}": "Vous avez reçu une nouvelle notification de {0} {1}",
   "You have unpaid invoices that will soon reach their limit date.": "Vous avez des factures impayées qui arrivent bientôt à échéance.",
   "You have {0} new comments on \"{1}\"": "Vous avez {0} nouveaux commentaires sur \"{1}\"",
   "You have {0} new replies to your comment": "Vous avez {0} nouvelles réponses à votre commentaire",
@@ -1501,7 +1501,7 @@
   "votes": "votes",
   "with project task link type": "avec le type de lien de tâche de projet",
   "z-A": "Z-A",
-  "{0} - Notifications from Goovee": "{0} - Notifications de Goovee",
+  "{0} - Notifications from {1}": "{0} - Notifications de {1}",
   "{0} added a comment": "{0} a ajouté un commentaire",
   "{0} added a comment on {1}": "{0} a ajouté un commentaire sur {1}",
   "{0} application(s) enabled": "{0} application(s) activée(s)",
